requirements.txt
torch>=2.1.0
torch-geometric>=2.4.0
numpy>=1.24.0
aiohttp>=3.9.0
openai>=1.12.0
tiktoken>=0.5.0
tenacity>=8.2.0
python-dotenv>=1.0.0
shortuuid>=1.0.11
class-registry>=2.1.2
sentence-transformers>=2.2.0
pyyaml>=6.0.1
tqdm>=4.66.0
datasets>=2.16.0
huggingface-hub>=0.20.0
jsonlines>=4.0.0
asyncio-pool>=0.6.0
